Rename and simplify useNotificationsSetup
